feat: add exchange logic

+++ currency_exchanger/exchange/dto/exchange_dto.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 from exchange.models import Currency
 
@@ -7,13 +8,13 @@ from exchange.models import Currency
 class ExchangeRateDto:
     base_currency: Currency
     target_currency: Currency
-    rate: float
-    amount: float
-    converted_amount: float
+    rate: Decimal
+    amount: Decimal
+    converted_amount: Decimal
 
 
 @dataclass
 class GetExchangeDto:
     from_currency_code: str
     to_currency_code: str
-    amount: float
+    amount: Decimal
